Retrieve resolutions only once on startup
Before the resolutions would always be retrieved when the options menu is opened. However, this caused a bit of lag while the menu was opening, which you could see. Now this lag is no longer there. Only downside is that new resolutions won't be visible if you swap your main monitor while the game is running.