Build with /MT intead of /MD
Out of nowhere many antiviruses started picking up on the /MD VC7 build. I tried re-compiling it with /MT and for some reason this yields no false positives at all. It's a bummer that I'm forced to switch over to /MT since it increases the file size for no good reason, but at least it gets rid of the false positives.